test/requirements-backup.txt
transformers
faiss-cpu
numpy
torch
requests
httpx==0.23.0
python-dotenv
pinecone-client
google-cloud-translate==2.0.1
google-auth
translate
groq
nltk
deepl
spacy